+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to Consolation will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 We adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- Option to use custom GUI skin (useful to customize scrollbar width)
+
 ## [1.3.1] - 2022-07-27
 
 ### Fixed